Fixed all complier warning on windows (using visual studio 2013) and updated .gitignore for windows.

+++ dbfdump.c
@@ -177,7 +177,7 @@ int main( int argc, char ** argv )
 
 	eType = DBFGetFieldInfo( hDBF, i, szTitle, &nWidth, &nDecimals );
 	if( (int) strlen(szTitle) > nWidth )
-	    panWidth[i] = strlen(szTitle);
+	    panWidth[i] = (int)strlen(szTitle);
 	else
 	    panWidth[i] = nWidth;
